+++ WebApplication1/Models/DTO/Comments.cs
@@ -12,7 +12,7 @@ namespace WebApplication1.Models.DTO
     {
         [Key]
         [Column("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [Column("body")]
         public string Body { get; set; }
@@ -21,9 +21,9 @@ namespace WebApplication1.Models.DTO
         public string Title { get; set; }
 
         [Column("post_id")]
-        public long PostId { get; set; }
+        public int PostId { get; set; }
 
         [Column("user_id")]
-        public long UserId { get; set; }
+        public int UserId { get; set; }
     }
 }
