+++ WebApplication1/Models/Front/Post.cs
@@ -10,7 +10,7 @@ namespace WebApplication1.Models.Front
         public string Title { get; set; }
         public string Body { get; set; }
         public string Description { get; set; }
-        public long CategoryId { get; set; }
-        public long UserId { get; set; }
+        public int CategoryId { get; set; }
+        public int UserId { get; set; }
     }
 }
